move gen release notes logic into a package
