+++ go.mod
@@ -4,8 +4,8 @@ go 1.17
 
 require (
 	github.com/google/go-github/v39 v39.2.0
-	github.com/rogpeppe/go-internal v1.3.0
 	github.com/sirupsen/logrus v1.8.1
+	golang.org/x/mod v0.3.0
 	golang.org/x/oauth2 v0.0.0-20211104180415-d3ed0bb246c8
 )
 
@@ -15,6 +15,7 @@ require (
 	golang.org/x/crypto v0.0.0-20210817164053-32db794688a5 // indirect
 	golang.org/x/net v0.0.0-20210226172049-e18ecbb05110 // indirect
 	golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
 	google.golang.org/appengine v1.6.7 // indirect
 	google.golang.org/protobuf v1.25.0 // indirect
 )
